refactor: 금액을 BigDecimal로 처리하고 Money로 래핑

+++ src/main/java/lotto/model/LottoSeller.java
@@ -1,10 +1,12 @@
 package lotto.model;
 
+import java.math.BigDecimal;
+
 public class LottoSeller {
     private final LottoSellingPolicy lottoSellingPolicy = new LottoSellingPolicy();
     private final LottosGenerator lottosGenerator = new LottosGenerator();
 
-    public Lottos sell(int purchasingMoney) {
+    public Lottos sell(BigDecimal purchasingMoney) {
         int countLottos = lottoSellingPolicy.calcuateLottoCount(purchasingMoney);
         return lottosGenerator.generate(countLottos);
     }
